Agregar módulos de manejo de archivos y utilidades, y mejorar la estructura del menú en el Gestor de Gastos
