+++ .gitignore
@@ -1,1 +1,2 @@
 movimientos.json
+__pycache__
